pyproject.toml
[project]
name = "atomagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.16",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain>=0.3.23",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.28",
    "openai>=1.72.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
